345 - Context settings moved outside of modal
